moved smoth_indicator to separated file
